chore: sanitize components
