+++ apps/meteor/client/components/GenericTable/index.ts
@@ -1,15 +1,8 @@
-import GenericTable from './GenericTable';
-import HeaderCell from './HeaderCell';
-
-export default Object.assign(GenericTable, {
-	HeaderCell,
-});
-
-export * from './V2/GenericTable';
-export * from './V2/GenericTableBody';
-export * from './V2/GenericTableCell';
-export * from './V2/GenericTableHeader';
-export * from './V2/GenericTableHeaderCell';
-export * from './V2/GenericTableLoadingRow';
-export * from './V2/GenericTableLoadingTable';
-export * from './V2/GenericTableRow';
+export * from './GenericTable';
+export * from './GenericTableBody';
+export * from './GenericTableCell';
+export * from './GenericTableHeader';
+export * from './GenericTableHeaderCell';
+export * from './GenericTableLoadingRow';
+export * from './GenericTableLoadingTable';
+export * from './GenericTableRow';
